feat: implement new layout components for Anilist search and user pages; refactor existing components for improved structure and readability
