+++ src/app/(protected)/anilist/_components/A_Toolbar.tsx
@@ -6,7 +6,7 @@ import { openModal } from '@/components/Modals';
 import ICON_SET from '@/constants/icons';
 import { AnilistMediaFilters } from '@/lib/types/anilist.types';
 
-const AnilistToolbar = ({
+const A_Toolbar = ({
     text,
     search,
     setSearch,
@@ -81,4 +81,4 @@ const AnilistToolbar = ({
     );
 };
 
-export default AnilistToolbar;
+export default A_Toolbar;
